pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weather-fusion"
version = "0.1.0"
description = "Dual-path ingestion pipeline for EHS-ready 10-day forecast emails"
authors = [{name = "John Cornelius"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "pandas>=2.2",
    "numpy>=1.26",
    "xarray>=2024.1",
    "cfgrib>=0.9.11.0",
    "eccodes>=1.6.1",
    "feedparser>=6.0",
    "beautifulsoup4>=4.12",
    "jinja2>=3.1",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "pytest-cov>=5.0",
    "freezegun>=1.4",
    "requests-mock>=1.11",
]

[project.scripts]
ehs_forecast = "weatherfusion.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
